refactor(veiculo): altera tipo do atributo
PL-3670

+++ src/main/java/br/com/zalf/prolog/webservice/v3/frota/veiculo/_model/VeiculoListagemDto.java
@@ -46,7 +46,7 @@ public class VeiculoListagemDto {
     String nomeModelo;
     @ApiModelProperty(value = "Código do diagrama do veículo", required = true, example = "1")
     @NotNull
-    Long codDiagrama;
+    Short codDiagrama;
     @ApiModelProperty(value = "Quantidade de eixos dianteiros", required = true, example = "1")
     @NotNull
     Long qtdEixosDianteiros;
